refactor(core): isolate viewport autoscaling internals
