+++ packages/core/src/scales/scaleResolution.js
@@ -13,7 +13,14 @@ import { configureDomain } from "../scale/scale.js";
 
 import ScaleInstanceManager from "./scaleInstanceManager.js";
 import { resolveScalePropsBase } from "./scalePropsResolver.js";
-import DomainPlanner, { isVisibleDomainRef } from "./domainPlanner.js";
+import DomainPlanner from "./domainPlanner.js";
+import {
+    getViewportConstraints,
+    getViewportDependencies,
+    isVisibleDomainRef,
+    isViewportDataReady,
+    VisibleDomainScheduler,
+} from "./visibleDomain.js";
 import ScaleInteractionController from "./scaleInteractionController.js";
 import { validateScaleTypeCompatibility } from "./scaleRules.js";
 import {
@@ -27,7 +34,6 @@ import {
 import { getAccessorDomainKey, isScaleAccessor } from "../encoder/accessor.js";
 import {
     getEncoderAccessors,
-    getEncoderDataAccessor,
     isSecondaryChannel,
     primaryPositionalChannels,
 } from "../encoder/encoder.js";
@@ -45,9 +51,6 @@ import {
     normalizeIntervalForSelection,
 } from "./selectionDomainUtils.js";
 import { toExternalIndexLikeInterval } from "./indexLikeDomainUtils.js";
-import { isSubtreeLazyReady } from "../view/dataReadiness.js";
-
-const VIEWPORT_AUTOSCALE_DEBOUNCE = 150;
 
 // Register scaleLocus to Vega-Scale.
 // Loci are discrete but the scale's domain can be adjusted in a continuous manner.
@@ -124,13 +127,8 @@ export default class ScaleResolution {
         range: new Set(),
     };
 
-    /** @type {(() => void)[]} */
-    #viewportDomainUnsubscribers = [];
-
-    /** @type {ReturnType<typeof setTimeout> | undefined} */
-    #viewportDomainTimer;
-
-    #viewportDomainWaitingForData = false;
+    /** @type {VisibleDomainScheduler} */
+    #visibleDomainScheduler;
 
     /** @type {ScaleInstanceManager} */
     #scaleManager;
@@ -198,6 +196,18 @@ export default class ScaleResolution {
             getType: () => this.type,
             getLocusExtent: (assembly) => this.#getLocusExtent(assembly),
             fromComplexInterval: this.fromComplexInterval.bind(this),
+        });
+
+        this.#visibleDomainScheduler = new VisibleDomainScheduler({
+            hasVisibleDomain: () => this.#domainAggregator.hasVisibleDomain(),
+            getDependencies: () =>
+                getViewportDependencies(this.#dataDomainMembers, this),
+            isReady: () =>
+                isViewportDataReady(
+                    this.#getActiveMembers(this.#dataDomainMembers),
+                    (member) => this.#getViewportConstraints(member)
+                ),
+            update: () => this.reconfigureDomain(),
         });
 
         this.#scaleManager = new ScaleInstanceManager({
@@ -568,7 +578,7 @@ export default class ScaleResolution {
         this.#invalidateConfiguredDomain();
         this.#refreshSelectionDomainParamSubscriptions();
         this.#refreshConfiguredDomainExprSubscriptions();
-        this.#refreshViewportDomainSubscriptions();
+        this.#visibleDomainScheduler.refresh();
     }
 
     #markMembersDirty() {
@@ -657,7 +667,7 @@ export default class ScaleResolution {
         this.#invalidateConfiguredDomain();
         this.#refreshSelectionDomainParamSubscriptions();
         this.#refreshConfiguredDomainExprSubscriptions();
-        this.#refreshViewportDomainSubscriptions();
+        this.#visibleDomainScheduler.refresh();
         this.#recreateInitializedScaleAndNotifyDomain();
     }
 
@@ -671,7 +681,7 @@ export default class ScaleResolution {
             this.#invalidateConfiguredDomain();
             this.#refreshSelectionDomainParamSubscriptions();
             this.#refreshConfiguredDomainExprSubscriptions();
-            this.#refreshViewportDomainSubscriptions();
+            this.#visibleDomainScheduler.refresh();
             this.#recreateInitializedScaleAndNotifyDomain();
         }
     }
@@ -717,60 +727,13 @@ export default class ScaleResolution {
      * @returns {{ channel: "x" | "y", domain: [number, number], accessor: import("../types/encoder.js").Accessor, accessor2?: import("../types/encoder.js").Accessor }[]}
      */
     #getViewportConstraints(member) {
-        /** @type {{ channel: "x" | "y", domain: [number, number], accessor: import("../types/encoder.js").Accessor, accessor2?: import("../types/encoder.js").Accessor }[]} */
-        const constraints = [];
-
-        for (const channel of primaryPositionalChannels) {
-            const resolution = member.view.getScaleResolution(channel);
-            if (!resolution || resolution === this) {
-                continue;
-            }
-            if (this.#hasVisibleDependencyPath(resolution, new Set())) {
-                throw new Error(
-                    `Viewport-derived scale domains form a dependency cycle in view "${member.view.getPathString()}".`
-                );
-            }
-
-            const scale = resolution.getScale();
-            if (!isContinuous(scale.type) || isDiscrete(scale.type)) {
-                continue;
-            }
-
-            const encoder = member.view.mark.encoders?.[channel];
-            if (!encoder) {
-                continue;
-            }
-            const accessor = getViewportAccessor(encoder);
-            if (!accessor) {
-                continue;
-            }
-
-            const secondaryChannel = channel === "x" ? "x2" : "y2";
-            const secondaryEncoder =
-                member.view.mark.encoders?.[secondaryChannel];
-            const accessor2 = secondaryEncoder
-                ? getViewportAccessor(secondaryEncoder)
-                : undefined;
-            const domain = resolution.getDomain();
-            constraints.push({
-                channel,
-                domain: [domain[0], domain.at(-1)],
-                accessor,
-                ...(accessor2 && { accessor2 }),
-            });
-        }
-
-        if (constraints.length === 0) {
-            const viewPath =
-                member.view.getPathString?.() ??
-                member.view.name ??
-                "(unknown)";
-            throw new Error(
-                `Viewport-derived ${this.channel} domain in view "${viewPath}" requires an independent continuous positional scale.`
-            );
-        }
-
-        return constraints;
+        return getViewportConstraints(
+            member,
+            this,
+            /** @type {ChannelWithScale} */ (this.channel),
+            (resolution) =>
+                this.#hasVisibleDependencyPath(resolution, new Set())
+        );
     }
 
     /**
@@ -831,7 +794,7 @@ export default class ScaleResolution {
     dispose() {
         this.#clearSelectionDomainParamSubscriptions();
         this.#clearConfiguredDomainExprSubscriptions();
-        this.#clearViewportDomainSubscriptions();
+        this.#visibleDomainScheduler.clear();
         this.#listeners.domain.clear();
         this.#listeners.range.clear();
         this.#scaleManager.dispose();
@@ -925,93 +888,6 @@ export default class ScaleResolution {
             const unsubscribe = expr.subscribe(listener);
             this.#configuredDomainExprUnsubscribers.push(unsubscribe);
         }
-    }
-
-    #clearViewportDomainSubscriptions() {
-        for (const unsubscribe of this.#viewportDomainUnsubscribers) {
-            unsubscribe();
-        }
-        this.#viewportDomainUnsubscribers = [];
-        clearTimeout(this.#viewportDomainTimer);
-        this.#viewportDomainTimer = undefined;
-        this.#viewportDomainWaitingForData = false;
-    }
-
-    #refreshViewportDomainSubscriptions() {
-        this.#clearViewportDomainSubscriptions();
-        if (!this.#domainAggregator.hasVisibleDomain()) {
-            return;
-        }
-
-        /** @type {Set<ScaleResolution>} */
-        const dependencies = new Set();
-        for (const member of this.#dataDomainMembers) {
-            for (const channel of primaryPositionalChannels) {
-                const resolution = member.view.getScaleResolution(channel);
-                if (resolution && resolution !== this) {
-                    dependencies.add(resolution);
-                }
-            }
-        }
-
-        const listener = () => this.#scheduleViewportDomainUpdate(false);
-        for (const resolution of dependencies) {
-            resolution.addEventListener("domain", listener);
-            this.#viewportDomainUnsubscribers.push(() =>
-                resolution.removeEventListener("domain", listener)
-            );
-        }
-    }
-
-    /**
-     * @param {boolean} collectorChanged
-     */
-    #scheduleViewportDomainUpdate(collectorChanged) {
-        if (
-            collectorChanged &&
-            this.#viewportDomainWaitingForData &&
-            this.#isViewportDataReady()
-        ) {
-            this.#viewportDomainWaitingForData = false;
-            this.reconfigureDomain();
-            return;
-        }
-
-        clearTimeout(this.#viewportDomainTimer);
-        this.#viewportDomainWaitingForData = false;
-        this.#viewportDomainTimer = setTimeout(() => {
-            this.#viewportDomainTimer = undefined;
-            if (this.#isViewportDataReady()) {
-                this.reconfigureDomain();
-            } else {
-                this.#viewportDomainWaitingForData = true;
-            }
-        }, VIEWPORT_AUTOSCALE_DEBOUNCE);
-    }
-
-    #isViewportDataReady() {
-        for (const member of this.#getActiveMembers(this.#dataDomainMembers)) {
-            const collector = member.view.getCollector();
-            if (!collector?.completed) {
-                return false;
-            }
-
-            /** @type {import("../data/sources/lazy/singleAxisLazySource.js").DataReadinessRequest} */
-            const request = {};
-            for (const constraint of this.#getViewportConstraints(member)) {
-                request[constraint.channel] = Array.from(constraint.domain);
-            }
-            if (
-                !isSubtreeLazyReady(
-                    member.view,
-                    request,
-                    (view) => view === member.view
-                )
-            ) {
-                return false;
-            }
-        }
-        return true;
     }
 
     #hasRenderedMember() {
@@ -1133,7 +1009,7 @@ export default class ScaleResolution {
                 this.#domainAggregator.hasVisibleDomain() &&
                 this.#initialDomainFinalized
             ) {
-                this.#scheduleViewportDomainUpdate(true);
+                this.#visibleDomainScheduler.schedule(true);
             } else {
                 this.reconfigureDomain();
             }
@@ -1945,17 +1821,6 @@ function intervalsEqual(a, b) {
     }
 
     return a.length === b.length && shallowArrayEquals(a, b);
-}
-
-/**
- * @param {import("../types/encoder.js").Encoder} encoder
- * @returns {import("../types/encoder.js").Accessor | undefined}
- */
-function getViewportAccessor(encoder) {
-    return (
-        getEncoderDataAccessor(encoder) ??
-        getEncoderAccessors(encoder).find(isScaleAccessor)
-    );
 }
 
 /**
